lectures/6.chailit with stream agent/chainlit_stream/pyproject.toml
[project]
name = "chainlit-stream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "asif-ml-developer", email = "asif.ml.developer@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.1",
    "openai-agents>=0.0.4",
    "python-dotenv>=1.0.1",
]

[project.scripts]
chainlit-stream = "chainlit_stream:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
